Fix vehicle access page typing
